調整後端為MVC架構
